+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 * `CounterZeroFilter`, `GaugeZeroFilter` and `TimerZeroFilter` has been added. They wrap monitors and only returns values to the poller when changes have been registered.
 * `MinMaxAvgGauge` has been added. It allows you to detect extreme values that would otherwise disappear in an average calculation.
 * `OkanshiMonitor.DefaultStep` has been deleted since it was forgotten in an earlier clean up process.
+* `OkanshiMonitor.PerformanceCounter` bugfix so it now returns an instance.
 * Unnecessary calculation has been removed from `Counter`, and locks are no longer used in the counter, instead atomics are used.
 * `Counter` now allows increment by negative numbers.